% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genius_album_fetcher.R
\name{genius_album_fetcher}
\alias{genius_album_fetcher}
\title{Fetch Album Tracklist and Lyrics}
\usage{
genius_album_fetcher(artists, albums)
}
\arguments{
\item{artists}{Character. The name of the artist, or a vector of artist names.}

\item{albums}{Character. The name of the album, or a vector of album names.
Each album in a vector much match the artist.}
}
\value{
A tibble with track titles, lyrics, and metadata for all specified artist
and album.
}
\description{
Retrieves album tracklist and lyrics for a specific artist and album from Genius.
Or the function accepts a vector of artist and a vector of corresponding album
to fetch data for multiple artist and album.
}
\examples{
# Fetch the tracklist and lyrics for Nirvana's album Nevermind
genius_album_fetcher("Nirvana", "Nevermind")
}
