% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genius_discography_fetcher.R
\name{genius_discography_fetcher}
\alias{genius_discography_fetcher}
\title{Fetch Discography for An Artist}
\usage{
genius_discography_fetcher(artist)
}
\arguments{
\item{artist}{Character. The name of the artist.}
}
\value{
A tibble with album titles and release year.
}
\description{
Retrieves an album list for a specific artist from Genius with the year of release
}
\examples{
# Fetch the discography for The Rolling Stones
genius_discography_fetcher("The Rolling Stones")
}
