% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genius_song_fetcher.R
\name{genius_song_fetcher}
\alias{genius_song_fetcher}
\title{Fetch Song Lyrics}
\usage{
genius_song_fetcher(artist, song_title)
}
\arguments{
\item{artist}{Character. The name of the artist}

\item{song_title}{Character. The name of the song}
}
\description{
Retrieves album tracklist and lyrics for a specific artist and album from Genius.
Or the function accepts a vector of artists and a vector of corresponding albums
to fetch data for multiple artists and albums.
}
\examples{
# Fetch the lyrics for The Rolling Stone's "Paint It Black"
genius_song_fetcher("The Rolling Stones", "Paint It Black")
}
