% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_page_fetcher.R
\name{webfetcheR}
\alias{webfetcheR}
\title{Fetch Text or Lyrics from a Web Page}
\usage{
webfetcheR(url)
}
\arguments{
\item{url}{Character. The url of the web page to scrape text from.

#' @examples
Fetch the text for Book 2 of the Iliad
webfetcheR("https://www.theoi.com/Text/HomerIliad2.html)}
}
\description{
Retrieves text or lyrics from a webpage other than Genius.
}
